+++ packages/dashed-date/index.js
@@ -1,10 +1,10 @@
 export {
   shiftDay, shiftMonth, shiftQuarter, shiftYear
-} from './src/shiftPeriod'
+} from './src/shiftPeriod.js'
 export {
   year, month, day, yearMonth, season,
   seasonEnds, seasonLo, seasonHi, seasonLoHi, monthHi, monthLo, monthLoHi
-} from './src/takePeriod'
+} from './src/takePeriod.js'
 export {
   within
-} from './src/withinPeriod'
+} from './src/withinPeriod.js'