+++ packages/date-shift/index.js
@@ -1,3 +1,3 @@
-export { within }                                                      from './src/withinPeriod'
-export { shiftDay, shiftMonth, shiftQuarter, shiftYear }               from './src/shiftPeriod'
-export { monthLo, monthHi, monthLoHi, seasonLo, seasonHi, seasonLoHi } from './src/takePeriod'
+export { within }                                                      from './src/withinPeriod.js'
+export { shiftDay, shiftMonth, shiftQuarter, shiftYear }               from './src/shiftPeriod.js'
+export { monthLo, monthHi, monthLoHi, seasonLo, seasonHi, seasonLoHi } from './src/takePeriod.js'