+++ packages/format-date-time/index.js
@@ -1,1 +1,1 @@
-export { FormatDateTime, formatDateTime, TIME_CONFIG, DATE_CONFIG } from './src/FormatDateTime'
+export { FormatDateTime, formatDateTime, TIME_CONFIG, DATE_CONFIG } from './src/FormatDateTime.js'