+++ packages/format-date/index.js
@@ -1,2 +1,2 @@
-export { DATE_CONFIG } from './src/dateConfig'
-export { FormatDate, formatDate } from './src/FormatDate'
+export { DATE_CONFIG } from './src/dateConfig.js'
+export { FormatDate, formatDate } from './src/FormatDate.js'