+++ packages/format-time/index.js
@@ -1,2 +1,2 @@
-export { FormatTime, formatTime } from './src/FormatTime'
-export { TIME_CONFIG } from './src/timeConfig'
+export { FormatTime, formatTime } from './src/FormatTime.js'
+export { TIME_CONFIG } from './src/timeConfig.js'