+++ packages/linger/index.js
@@ -1,5 +1,5 @@
-export { linger }   from './src/linger'
-export { Escape }   from './src/Escape'
-export { Cylinder } from './src/Cylinder'
+export { linger }   from './src/linger.js'
+export { Escape }   from './src/Escape.js'
+export { Cylinder } from './src/Cylinder.js'