+++ packages/timeout/index.js
@@ -1,2 +1,2 @@
-export { timeout }                  from './src/timeout'
-export { intime, ontime, overtime } from './src/timings'
+export { timeout }                  from './src/timeout.js'
+export { intime, ontime, overtime } from './src/timings.js'