+++ packages/timestamp/index.js
@@ -1,4 +1,4 @@
 export {
   year, date, roughTime, time, dateTime,
   today, roughly, roughlyNow, now, present
-} from './src/methods'
+} from './src/methods.js'