+++ packages/util-leap-year/index.js
@@ -1,1 +1,1 @@
-export { isLeap } from './src/utils'
+export { isLeap } from './src/utils.js'