+++ packages/util-month-days/index.js
@@ -1,3 +1,3 @@
-export { monthDays }          from './src/monthDays'
-export { bigMonth }           from './src/bigMonth'
-export { seasonLast, monthToSeason } from './src/seasonLast'
+export { monthDays }          from './src/monthDays.js'
+export { bigMonth }           from './src/bigMonth.js'
+export { seasonLast, monthToSeason } from './src/seasonLast.js'