+++ packages/util-ymd/index.js
@@ -1,2 +1,2 @@
-export { prevMonth, nextMonth } from './src/shiftMonth'
-export { calibre }              from './src/calibre'
+export { prevMonth, nextMonth } from './src/shiftMonth.js'
+export { calibre }              from './src/calibre.js'