+++ packages/dashed-date/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @valjoux/dashed-date
 
+## 0.4.2
+
+### Patch Changes
+
+- minor fix
+- Updated dependencies
+  - @valjoux/convert@0.4.2
+  - @valjoux/date-shift@0.4.2
+  - @valjoux/util-leap-year@0.4.2
+  - @valjoux/util-month-days@0.4.2
+
 ## 0.4.1
 
 ### Patch Changes