+++ packages/date-shift/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @valjoux/date-shift
 
+## 0.4.2
+
+### Patch Changes
+
+- minor fix
+- Updated dependencies
+  - @valjoux/convert@0.4.2
+  - @valjoux/day-shift@0.4.2
+  - @valjoux/util-month-days@0.4.2
+  - @valjoux/util-ymd@0.4.2
+
 ## 0.4.1
 
 ### Patch Changes