+++ packages/format-date/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @valjoux/format-date
 
+## 0.4.2
+
+### Patch Changes
+
+- minor fix
+
 ## 0.4.1
 
 ### Patch Changes