+++ packages/linger/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @valjoux/linger
 
+## 0.4.2
+
+### Patch Changes
+
+- minor fix
+- Updated dependencies
+  - @valjoux/timeout@0.4.2
+
 ## 0.4.1
 
 ### Patch Changes