+++ packages/util-bitwise/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @valjoux/util-bitwise
 
+## 0.4.2
+
+### Patch Changes
+
+- minor fix
+
 ## 0.4.1
 
 ### Patch Changes